# reduce webpack build time by ~50s