--- a/webpack.config.js
+++ b/webpack.config.js
@@ -105,8 +105,11 @@ function production () {  // eslint-disable-line
   prod.plugins.push(new webpack.optimize.OccurrenceOrderPlugin(true))
   if (env !== 'test') {
     prod.plugins.push(new UglifyJsPlugin({
+      parallel: true,
       uglifyOptions: {
         compress: false,
+        // skip pre-minified libs
+        exclude: [/\.min\.js$/gi],
         mangle: {
           reserved: ['module', 'exports', 'require']
         }