@evq 2017-10-19
It makes sense that compression negligibly affects the overall build size, I thought we had it enabled for a possible speedup on load. thinking on this though, it's not like we're moving things over the network - so it's probably slowing things down there as well to decompress