@cezaraugusto 2017-10-23
@evq sorry I don't get what you mean. Which scenario should I test to ensure your concern is addressed?