@bbondy 2017-10-23
I'm a bit concerned that this will reduce perf for a benefit in build time.