@bsclifton 2017-11-02
@cezaraugusto I think @evq is saying that this should increase performance for loading the JavaScript, since it will no longer need to do decompression before processing

Is there any impact to startup time? (after installing a packaged build).  The time spent building a package doesn't matter too much, since we automate that w/ Jenkins (the time savings there wouldn't win much).