@bsclifton 2017-11-09
Closing for now- once we can address the items Bondy has mentioned, maybe let's re-open and we can re-review 😄