@bsclifton 2017-12-04
Heads up: this PR sped up build time by ~45 seconds on my machine while keeping minification 😄 
https://github.com/brave/browser-laptop/pull/12129