@cezaraugusto 2017-12-04
ya `parallel: true` is a great helper